10242: Move into Forms namespace
